requirements.txt
# Requirements for Open Source Scout

# Web UI
streamlit>=1.30.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# PDF Generation
reportlab>=4.0.0

# Git Operations
gitpython>=3.1.40

# Retry Logic
tenacity>=8.2.0

# Markdown Processing
markdown>=3.5.0

# Testing
pytest>=7.4.0
